osm-controller: use fixed port number for ADS server
Resolves #2190 by using a constant port number for the
ADS server in osm-controller. The port number was partially
configurable via the controller's command line flags but
this isn't enough to make the port number configurable.
Since there is no use case for this port to be configurable,
use a fixed port number. Also, renames the name of the port
for clarity.

Resolves #2190

Signed-off-by: Shashank Ram <shashr2204@gmail.com>