ref(pkg/injector): use XDS structs for envoy bootstrap config
Updates the envoy bootstrap config to be constucted using XDS
structs directly instead of maps. The XDS objects are marshalled
to JSON and then to YAML. Using the XDS structs directly will help
ensure the objects are configured correctly and all required
fields are set. Additionally, this change will make it easier to
to enforce the use of the same api version across the injector and
controller.

Signed-off-by: Jackie Elliott <jaellio@microsoft.com>