ref(docs): use relative links to repo files
When all of the docs in this repo were published to
https://docs.openservicemesh.io, we decided to use absolute links to
other files in this repo so they would work from the website. Now that
these docs are no longer rendered on the website, these links only need
to work on GitHub, which supports relative links. Making the links
relative unambiguously refers to the file in the same commit and
eliminates any need to manually keep them up to date when a new release
branch is created.

Signed-off-by: Jon Huhn <johuhn@microsoft.com>