[testing] Do not send SDS on full config updates
Theres a misshandle of XDS resources as our full config push could record we sent some
resources that could be ignored by envoy if envoy is still not interested in them.

Later requests from envoy for that resource and that nonce could ask for resources we have
already sent but were ignored at a previous time.

This commit prevents pushing certificates on full config updates, which will make SDS almost
entirely envoy-driven; it will prevent us pushing unnecessary certificates that envoy did not
request, hence never allowing us to send pootentially ignored resources at a response.

Certificate rotation can still cause issues if a MakeAllSecrets is called and a new certificate that
was not present in resources is added during that time (say, a config change happened while cert rotation
also happened). An issue will be opened for this and we will fix this separately.

Signed-off-by: Eduard Serra <eduser25@gmail.com>