envoy: add additional metadata for the proxy's kind
Adds additional metadata in the proxy's XDS cert
to identify the proxy's kind.

Currently, osm-controller only supports programming proxies
that run as a sidecar on app pods. As a part of the effort
to implement a proxy gateway, osm-controller will also
require to identify and program a proxy gateway. The
config for a gateway proxy will be completely different
compared to that of a sidecar proxy, and this change will
allow identifying the kind of the proxy to do so.

Existing code and tests have been updated to set the
proxy's kind to `sidecar`.

Part of #3501

Signed-off-by: Shashank Ram <shashr2204@gmail.com>