XDS Server: introducing snapshot cache
Go-control-plane includes an XDS server implementation that is interacted
through a map-like mechanisms that decouples generating proxy configuration
from its distribution.

https://github.com/envoyproxy/go-control-plane/blob/main/docs/cache/Snapshot.md

This commit adds support for initializing OSM control plane to use such mechanism, behind
a feature flag.

Using the cache currently suffers from some problems (Some work, Some are flacky, most will fail).
A lot more work is to be added on top of this implementation to debug these issues.

Meaningful changes:
- New featureflag added: `enable-snapshotcache`
- `server.go` will initialize the XDS server with or without the cache when enabled.
- Some changes were made to `ads/response.go` to accomodate both modes working.
Namely, it will now `get` the resources for a vertical, and later decide what to do with
them (either send to a connected proxy, or store them in the cache, depending if the cache
is enabled or not).
- `stream.go` was before the routine receiving events and triggering envoy re-configs. When
working on snapshot-cache, such routines do not exist, so a new routine to mimic the same
current design and behavior has been added (`snapshot_cache_stream.go`). It's purpose is to
listen to broadcast events from dispatcher, and issue a reconfiguration for all proxies when
a broadcast update is sent from dispatcher.

TODOs:
- Proxy objects are not created anymore in snapshot-cache based. Follow up discussion if
we want to have them or not is to happen.
- Zerolog doesn't implement Golang's logging interface and can't be used/passed to log
cache internals.
- Additional context validation (certificate) can happen on cache callback context.
- We have no way to infer the `kind` of proxy from a Pod. Right now it's temporarily
hardcoded as `sidecar`.

Signed-off-by: Eduard Serra <eduser25@gmail.com>