Fix autorebase label (#3545)
Removes the bot_account field to allow the autorebase feature to work. Originally, the feature (to specify the account being used to do the rebasing under) was free, but now it is a subscription-only feature. Now the autorebase feature will use any OSM account in the Mergify dashboard to perform the rebasing, which is not an issue because there's only one OSM account in Mergify.
Signed-off-by: Shalier Xia <shalierxia@microsoft.com>