feat(pkg/*): Lay some groundwork for multi cluster gateways.
Multi cluster gateway configurations will have multiple routes
with the same service name. This PR introduces a cluster field
to the MeshService, which currently always returns true for local.
We also change TrafficPolicy names to *always* include the cluster id
and the namespace, export a few functions, and introduce a Locality
type to  determine if a service is being reached from which locality.

Signed-off-by: Sean Teeling <seanteeling@microsoft.com>