fix(demo): remove host headers in SMI policies
The `:authority` header sent by the bookbuyer to the bookstore is
actually `bookstore.bookstore:14001` where the bookstore's
HTTPRouteGroup used in the manual demo specified a `host` header match
of `bookstore.bookstore`, resulting in

    "response_code_details": "route_not_found"

in the bookstore's logs for requests from the bookbuyer.

This change removes the `host` header matches all of the demo manifests.

Fixes #3616

Signed-off-by: Jon Huhn <johuhn@microsoft.com>