fix(pkg/envoy/cds): configure egress Envoy cluster for prometheus
Formats Envoy clusters created based on the applied egress
policies so that egress traffic can be observed with prometheus.
The change only applies to egress traffic enabled using egress
policies, not globally enabled egress. The AltStatName of the
cluster is formatted so that the stats for the cluster are grouped
with a metric we are already scraping for. Prior to this change,
the egress cluster name would be interpretted incorrectly due to
periods and colons and the stats would be grouped under a unique
metric. We were not scraping for this metric and it would be
difficult for customers to interpret.

Signed-off-by: jaellio <jaellio@microsoft.com>