envoy: encode proxy identity in XDS cert
This change refactors the code to encode the
proxy's identity in its XDS cert, which
can be used to derive its service identity
that is encoded in certs for proxy->proxy
communication.

The goal of this change is to avoid leaking
k8s service account across the XDS code, which
abstracts the proxy identity using the identity
pkg. As a part of this, unnecessary conversions
from CN -> service-account -> service-identity
are avoided when invoking APIs on the MeshCataloger
interface. Within XDS, service-identity ->
service-account conversions are only needed in
situations where k8s code has already leaked,
such as when inspecting the associated pod's spec.

Also renames some functions for clarity.

Part of #3168

Signed-off-by: Shashank Ram <shashr2204@gmail.com>