injector/iptables: resolve unnecessary TODO comment
Removes an unnecessary TODO comment carried
over from the initial init-iptables.sh script.
This TODO was meant to address a case where
a client sends traffic to itself using the
service IP, which would get routed back to
itself, and if such traffic needed to be
bypassed then an additional iptables rule
would be necessary. Since we do not have
any requirement to bypass additional traffic,
this change removes the ambiguous TODO comment.

Resolves #1266

Signed-off-by: Shashank Ram <shashr2204@gmail.com>