envoy/lds: refactor http connection manager + wasm filter
The existing code to build the HTTP connection
manager is unnecessarily complex and complicated
to test. This change refactor the code such
that it is easier to comprehend and test.

Additionally, removes the WASM stats headers for ingress
so that internal mesh details (pod name, namespace,
deployments etc.) are not leaked as a part of the
HTTP response headers.

- Removes mesh WASM config for ingress
- WASM, external auth, tracing related code
  have been consolidated into their respective
  files.
- HTTP connection manager is now built from its
  input config as opposed to passing the Configurator
  interface around multiple times.
- Fixes incorrect test expectations when registering
  mock expectations for function calls.
- Only adds the necessary WASM filters when `statsHeaders`
  are configured on the `lisenerBuilder` type.
- Renames tests to match the function names and moves
  updated tests to their corresponding test files.

Signed-off-by: Shashank Ram <shashr2204@gmail.com>