envoy: remove unnecessary peer validation cert for TLS (non-mTLS)
This change fixes an existing bug where when mTLS is disabled,
the root certificate is still being configured to validate the
client cert. In TLS mode (non-mTLS), the certificate presented
by the client should not be validated and hence there is no
reason to publish this peer cert validation context along with
the server cert.

A non-mTLS root certificate should never be requested by the proxy.

Required by #3779 to enable TLS based ingress to backends.

Signed-off-by: Shashank Ram <shashr2204@gmail.com>