contributing: allow squash & merge option
We have already been allowing the squash & merge
option to maintainers and contributors via
the github merge option or using the automerge-squash
label. This has been working fine so document this
as an option available to users.

Previously there was a concern on the original author's
signature being open to modification, but this is not
a problem given the DCO check ensures the commit was
created with a valid signature. Maintainers only
need to ensure they preserve the signatures of the
original authors when squashing and merging.

Signed-off-by: Shashank Ram <shashr2204@gmail.com>