fix(wasm): move metrics handles to RootContext
This change moves the metric definitions for the stats WASM module from
the per-stream `Context` to the persistent `RootContext` to prevent
requests from failing after 1-2 hours in the demo with
`response_code_details` of `wasm_fail_stream`.

I'm not 100% certain this is **the** solution, but I've been running a
cluster with this change and requests have been succeeding for 16+
hours.

Fixes #3958

Signed-off-by: Jon Huhn <johuhn@microsoft.com>