envoy: update version to v1.19.0
Updates the envoy version to v1.19.0 for both
OSM's sidecar and Contour's edge proxy.
Contour 1.18 uses Envoy v1.19.0 by default, but
the Bitnami Helm charts weren't updated to reflect
this change yet. This change also overrides Contour's
envoy image to match the default upstream version so
that the same version of envoy is used for both the
sidecar and ingress gateway.

Signed-off-by: Shashank Ram <shashr2204@gmail.com>