ingress: don't error if candidate group version is not available
If a candidate API GroupVersion (ex. networking.k8s.io/v1beta1) is not
available, don't error. Only error if none of the candidates are
available (already the case).
This is required for OSM to work in k8s v1.22, where the
networking.k8s.io/v1beta1 is deprecated.

Signed-off-by: Shashank Ram <shashr2204@gmail.com>