pkg/smi: allow filter options while listing items in cache
There is a use case to only set a subset of items in the
cache for TrafficTarget and TrafficSplit resources based
on certain fields being set on the resources. This change
implements optional filters that clients can use to filter
the returned items from the cache without needing to define
additional functions. With this change, the TrafficTarget
cache can be filtered based on the destination identity,
and the TrafficSplit cache can be filtered based on the
apex and backend service.

Required for #4052

Signed-off-by: Shashank Ram <shashr2204@gmail.com>