pkg/*: simplify interface function signatures (#4075)
This change simplifies function signatures when a slice
is being returned to the caller by letting the caller
of the function decide on error conditions. In a lot
of cases, nil error values are always being returned
by many functions, which results in duplicate errors
being propagated and logged across the call stack in
some cases. Removing unnecessary error paths simplifies
the code and makes it easier to test.

Signed-off-by: Shashank Ram <shashr2204@gmail.com>