envoy/eds: respond to EDS request for cluster with no endpoints
It's possible for a cluster created via CDS to have no endpoints,
such as an apex service referenced in a traffic split resource
that is not backed by any pods. Respond to the EDS request
for such a cluster with no endpoints so that Envoy does
not keep requesting for it.

Signed-off-by: Shashank Ram <shashr2204@gmail.com>