envoy/eds: respond with endpoints for all services
When generating the EDS response for all upstream
endpoints that the downstream proxy is allowed to
access, create an EDS response for a service without
endpoints as well. This is similar to 78455da0 which
allowed the XDS state machine to converge when there
are CDS clusters that use EDS for endpoint discovery.

Signed-off-by: Shashank Ram <shashr2204@gmail.com>