envoy: pin images to sha256 digest and update PullPolicy
Updates the default envoy image to use image digests
instead of tags to uniquely and immutably identify
a container image. Tags are mutable and are vulnerable
to supply chain attacks, so this change relies on image
digests to avoid such problems.

Additionally, updates the envoy container's image pull
policy to only pull the image if not present, as there
is no need to pull the image if it's already present.

Resolves #3715

Signed-off-by: Shashank Ram <shashr2204@gmail.com>