Makefile: add common target for control plane
Adds a common `make` target to build the control
plane that is used by both the linux and windows
build targets.
Also fixes an issue with the VERIFY_TAGS value being
overriden and adds an additional target to publish
the control plane image digests to a file. The image
digests will be made available as a part of a pre-release
workflow to publish the control plane images, and the
digests will be used to encode the default images in
the Helm chart and CLI.

Signed-off-by: Shashank Ram <shashr2204@gmail.com>