ref(build): build windows images with explicit buildx builder
This change modifies the Makefile targets for building Windows images
with buildx to use the `img-builder` builder explicitly because the
effects of the `--use` flag on `docker buildx create` will not persist
if a user changes the default builder. This way we still ensure a
builder that can build windows/amd64 images exists and use that each
time no matter what the default builder is.

Signed-off-by: Jon Huhn <johuhn@microsoft.com>