feat(injector): Set probe timeouts based on pod deployment spec
Fixes #4137

Overall we aim to maintain the following invariance:

For all health probes we want the following to be true:

All Envoy (implicit or explicit) timeouts >= the timeout specified in
the pod deployment spec.

Changes:

* Remove connect timeouts from all clusters
* For health probe clusters we set the route timeout to be equal to the
timeout set on the kubernetes pod. This is only applied on HTTP routes
since plain tcp routes have infinite timeout.

Signed-off-by: Sotiris Nanopoulos <sonanopo@microsoft.com>