fix(release): set version and build date for release builds
Defines VERSION and BUILD_DATE environment variables in the
pre-release.yaml. When the osm build occurs for a release the
VERSION and BUILD_DATE variables in the makefile will be
populated with the values of the env vars rather than the default
value of dev for VERSION and nothing for BUILD_DATE. This change
fixs the incorrect VERSION and BUILD_DATE logged in the osm control
plane components.

Signed-off-by: jaellio <jaellio@microsoft.com>
Signed-off-by: Sneha Chhabria <snchh@microsoft.com>