catalog/ingress: check backend's port in addition to name
Checks that the IngressBackend's port matches the TargetPort
of the service. This is important in deployments where there
are multiple MeshServices with the same name, in which case
the IngressBackend configuration corresponding to the service
name and port must be used. Multiple MeshService objects with
the same name and different port number will exist when there
is a k8s service with multiple ports per service.

Without this change, IngressBackend configuration for a service
with multiple ports can result in conflicts where duplicate
configs are generated per MeshService.

Signed-off-by: Shashank Ram <shashr2204@gmail.com>