go/deps: update Helm to v3.7.0 and kubernetes packages
Updates Helm to v3.7.0 which is a feature release. The previous
Helm version, v3.6.1 was dependent on a version of containerd that
required a bug fix. Helm needed to be updated to a version that
included the fixed containerd pkg. As a result, the kubernetes
packages are updated to v0.22.1. Also includes codchanges
resulting from upgrading the pkg versions.

Signed-off-by: jaellio <jaellio@microsoft.com>