envoy/registry: do not track disconnected proxies
Proxies are expected to disconnect when the corresponding
pods are recycled.
Since the cluster can go through a lot of churn, keeping
track of disconnected proxies results in additional
memory consumption and bookkeeping which is better avoided.
It is easy enough to know how many times a proxy has
reconnected to a control plane after a disconnect using
the osm_proxy_reconnect_count metric.

Signed-off-by: Shashank Ram <shashr2204@gmail.com>