fix(ads): properly sync proxy disconnect
This change fixes a goroutine leak leading to high memory usage when
Pods with proxies are being churned. The `StreamAggregatedResources()`
function launches a new goroutine running `receive()` whenever a new
proxy connects to osm-controller. Proxies may disconnect in a few
different scenarios, but from logs the most common case seemed to be
on `ctx.Done()` from inside `StreamAggregatedResources()`. When that
condition was hit however, nothing signalled the `receive()` goroutine
to stop, so those would run forever even after a proxy disconnected.

This change moves listening for `ctx.Done()` to `receive()` and relies
on `StreamAggregatedResources()` to exit when `receive()` finishes like
it was doing before.

Related to #4206

Signed-off-by: Jon Huhn <johuhn@microsoft.com>
Signed-off-by: Sneha Chhabria <snchh@microsoft.com>