messaging/broker: batch proxy update events
Implements batch processing of proxy update events
to reduce the number events propagated to the XDS
routine, through event coalescing, when the events
are triggered in close proximity in time. Without
this change, when a large number of events are
received in a short time interval, not coalescing
the events will trigger a large number of proxy
update events on the pub-sub, that will result
in the osm-controller's CPU spiking, and also
that of envoy based on how soon the XDS state
machine converges.

This implements a sliding window much like how
the dispatcher routine in mesh-catalog would
handle proxy events prior to the introduction
of a message broker, except that it resolves
potential memory leaks that might arise when
the timers are reset during a processing window.

It also updates a test to not use rate limiting
because it slows the test by about ~25s, and the
ticker test to use the counter exposed by the
broker.

Resolves #4238

Signed-off-by: Shashank Ram <shashr2204@gmail.com>