feat(metrics): monitored namespaces prometheus metric
Adds a monitored namespace counter guage. StartNamespaceCounter
is a goroutine that subscribes to the namespace added and deleted
events on the KubeEventPubSub. The metric is updated based on
these events.

The Monitored Namespace Count panel was updated to use the
MonitoredNamespaceCouter metric. The count is no longer calculated
by subracting the number of namespace deleted events from the
number of namespace added events obtained from the
k8sApiEventCounter. As descibed in #3918, the previous PromQL query
did not accurately determine the difference between the counter
vectors. This change removes the reliance on the k8sApiEventCounter
to determine resource counts.

Removes Number of Pods, and Services in the Mesh panels from the
Grafana dashboard. Currently, osm Prometheus metrics do not exist
to accurately determine the number of resources in the mesh.
Future PRs will add these metrics and add panels to the dashboard.

Addresses #3918

Signed-off-by: jaellio <jaellio@microsoft.com>