ingress/client: remove unnecessary check for namespace
Removes a redundant check in the ingress cache that checks
if the namespace is monitored. The check is reduntant because
there is already a check that ensures the ingress resource
namespace is the same as the MeshService namespace, where
MeshService's namespace is guaranteed to be monitored.

IsMonitoredNamespace() usage should be avoided if possible,
since repeated cache access of the backing thread-safe map
results in performance issues due to contention.

Signed-off-by: Shashank Ram <shashr2204@gmail.com>