fix(MeshConfig): Remove omitEmpty from bool values in the meshconfig
By default the meshConfig was ingnoring empty bool values which includes
false, hence a number of fields weren't showing up. This PR removes the
omitEmpty from all bool values in the meshConfig.

This is necessary because we have a number of flags set to false by default and they do not show up in the meshConfig, as they are interpreted as nil and skipped during JSON encoding when the omitempty tag is preset, making it difficult for users to decipher what a the value of a particular key is in the meshConfig

Signed-off-by: Sneha Chhabria <snchh@microsoft.com>