ref(build): remove need for helm dependency update
This change commits the subcharts in charts/osm/charts so `helm
dependency update` is no longer needed frequently since that command has
been failing recently in CI intermittently. `helm dependency update`
will still need to be run in the future only when a new version of a
subchart is needed.

Fixes #4284

Signed-off-by: Jon Huhn <johuhn@microsoft.com>