catalog: allow root service selector to match backend pods
Allows root service in a traffic split to have a selector
that matches the pods backing the leaf services. This
scenario was not tested before.

The change updates the inbound policy generation code
in a way such that policies are built for each upstream
service (including root service), where the upstream
service is guaranteed to be unique. Additionally,
traffic directed to the root service are routed to
a cluster corresponding to the root service and not
the leaf service.

Resolves #4296

Signed-off-by: Shashank Ram <shashr2204@gmail.com>