ref(tcp): support tcp-server-first inside the mesh
This change adds the capability for OSM to handle ports with
`appProtocol: tcp-server-first` on services within the mesh.
Specifically, the main change is to add handling for `tcp-server-first`
ports in the mesh when we calculate the predicate function used to
disable the TLS and HTTP inspector listener filters. Previously, this
was only done for `tcp-server-first` ports specified in an Egress
policy resource.

Fixes #4313

Signed-off-by: Jon Huhn <johuhn@microsoft.com>