docs: add release notes doc
Adds a release notes doc to track breaking
changes and deprecated functionality. Going
forward, this doc must be updated in the same
PR that introduces such changes.

This is required to correctly track such changes
and document an upgrade path to users based on
the changes that are not backward compatible.

Signed-off-by: Shashank Ram <shashr2204@gmail.com>