fix(cli): set sidecar injection annotation to disabled
Modifies the patch command on `osm namespace add --disable-sidecar
-injection` to set the sidecar injection annotation to "disabled"
instead of null. Patching the annotation key with a null value
should have removed the annotation if it existed or had no effect
if the annotation key did not exist. Currently, when the
sidecar injection annotation does not exist on the namespace, the
patch mentioned above adds te annotation with a value of an empty
string. This result is errors in the osm-injector when pods are
added to the namespace.

This PR also includes a change to notify the user if the namespace
being added is already monitored by the mesh.

Signed-off-by: jaellio <jaellio@microsoft.com>