feat(cli): add remote version to osm version output
Adds remote version to osm version CLI command output. By default
the CLI and Mesh version will be included in the output. A remote
flag (by default set to true) can be set to false to only output
the CLI version. If there is any failure getting the mesh version,
the CLI version is printed along with the failure message.

Part of #4289

Signed-off-by: jaellio <jaellio@microsoft.com>