fix(grafana): remove hardcoded control plane namespace (#4454)
Removes the hardcoded default control plane namespace from the
Mesh and Envoy Details dashboard. Parameterizes the control plane
namespace. When installing OSM with automatic provisioning of
prometheus and grafana the control plane namespace variable will
be set to the namespace where OSM is being installed. When
importing the Mesh and Envoy Details dashboard, the user will have
the option to set the osm control plane namespace just as they are
able to set the datasource.

Signed-off-by: jaellio <jaellio@microsoft.com>