tests/framework: wait for deletion and dump namespace
Waits for resources created by Helm to delete prior
to performing deletion of the entire namespace.
Without this, some dangling resources can be left
behind in the cluster such and the namespace could
be stuck in the terminating state.
Also dumps the namespaces that fail to terminate. This
is necessary to root cause namespace deletion issues.

Signed-off-by: Shashank Ram <shashr2204@gmail.com>