configurator: simplify MeshConfig API
Simplifies the signature of `GetMeshConfig()` in
the interface to return an object instead of pointer.
MeshConfig is guaranteed to exist upon controller
start and there is already a fallback to return
a default MeshConfig object if not found in the
cache. Returning a pointer results in unnecessary
nil pointer handling on the caller side, which
makes the usage of this function indeterministic
when the implementation is deterministic.
This change guarantees to the caller of this function
that they will always receive a valid MeshConfig to
operate on.

Signed-off-by: Shashank Ram <shashr2204@gmail.com>