feat(viz): control-plane timings viz improvements
This PR changes Grafana visualization so that
injector timings and cert issuance timings are properly
visualized in the heatmap views.

This PR also adds a distribution histogram view of
the 2 timings as discussed with project maintainers.
This is because the aforementioned heatmap shows
a "histogram time series" view of the timing distributions
per minute, while the 2 new histogram distribution panels
show the "total distribution of all timings over all time".

Resolves #4452.

Signed-off-by: Johnson Shi <Johnson.Shi@microsoft.com>