fix(injector): make init container pullPolicy configurable
This change uses the `osm.image.pullPolicy` value in the Helm chart for
the pull policy configured for injected init containers to make the
behavior consistent with other OSM images. Previously, using
`osm.image.pullPolicy=IfNotPresent` and `osm.image.tag=latest` would
configure the init container with that tag but an empty pull policy
which would be implicitly `Always`, which fails in environments where
the image is not meant to be pulled.

Fixes #3809

Signed-off-by: Jon Huhn <johuhn@microsoft.com>