fix(proxy): remove proxy on "osm namespace remove"
This fixes the issue where running "osm namespace remove"
on a namespace results in the namespace being removed,
but not the pods within the namespace. The Envoy certificates
are still not released, the pods (and their proxies) are still
considered part of the mesh, and the connected-proxies metric
count is not properly decremented.

Resolves #4166.

Signed-off-by: Johnson Shi <Johnson.Shi@microsoft.com>